+++ lib/Property/ICalendar/DateTime.php
@@ -3,7 +3,9 @@
 namespace Sabre\VObject\Property\ICalendar;
 
 use DateTimeInterface;
+use DateTimeZone;
 use Sabre\VObject\Property;
+use Sabre\VObject\Parser\MimeDir;
 use Sabre\VObject\DateTimeParser;
 use Sabre\VObject\TimeZoneUtil;
 
@@ -116,11 +118,16 @@ class DateTime extends Property {
      * first will be returned. To get an array with multiple values, call
      * getDateTimes.
      *
+     * If no timezone information is known, because it's either an all-day
+     * property or floating time, we will use the DateTimeZone argument to
+     * figure out the exact date.
+     *
+     * @param DateTimeZone $timeZone
      * @return DateTimeImmutable
      */
-    function getDateTime() {
+    function getDateTime(DateTimeZone $timeZone = null) {
 
-        $dt = $this->getDateTimes();
+        $dt = $this->getDateTimes($timeZone);
         if (!$dt) return null;
 
         return $dt[0];
@@ -130,20 +137,26 @@ class DateTime extends Property {
     /**
      * Returns multiple date-time values.
      *
+     * If no timezone information is known, because it's either an all-day
+     * property or floating time, we will use the DateTimeZone argument to
+     * figure out the exact date.
+     *
+     * @param DateTimeZone $timeZone
      * @return DateTimeImmutable[]
+     * @return \DateTime[]
      */
-    function getDateTimes() {
+    function getDateTimes(DateTimeZone $timeZone = null) {
 
-        // Finding the timezone.
-        $tz = $this['TZID'];
+        // Does the property have a TZID?
+        $tzid = $this['TZID'];
 
-        if ($tz) {
-            $tz = TimeZoneUtil::getTimeZone((string)$tz, $this->root);
+        if ($tzid) {
+            $timeZone = TimeZoneUtil::getTimeZone((string)$tzid, $this->root);
         }
 
         $dts = [];
         foreach($this->getParts() as $part) {
-            $dts[] = DateTimeParser::parse($part, $tz);
+            $dts[] = DateTimeParser::parse($part, $timeZone);
         }
         return $dts;